// src/types/enums/order/order.ts
export enum ORDER_STATUS_ENUM {
    PAYMENT_CONFIRMED = "PAYMENT_CONFIRMED",
    WAITING_FOR_PAYMENT = "WAITING_FOR_PAYMENT",
    WAITING_FOR_CONFIRMATION = "WAITING_FOR_CONFIRMATION",
    INITIALIZED_FOR_PAYMENT = "INITIALIZED_FOR_PAYMENT",
    PROCESSING = "PROCESSING",
    SENT = "SENT",
    CANCELED = "CANCELED",
    CANCELED_PAYMENT_TIMEOUT = "CANCELED_PAYMENT_TIMEOUT",
    REFUNDED = "REFUNDED",
    IN_CART = "IN_CART",
}